reorganize StepExecution.java inherited classes
