+++ src/main/java/executor/service/service/stepexecutors/StepExecution.java
@@ -1,0 +1,12 @@
+package executor.service.service.stepexecutors;
+
+import executor.service.model.entity.Step;
+import org.openqa.selenium.WebDriver;
+
+public interface StepExecution {
+
+  String getStepAction();
+
+  void step(WebDriver webDriver, Step step);
+
+}
