+++ src/main/java/executor/service/service/stepexecutors/StepExecutionFabric.java
@@ -1,0 +1,10 @@
+package executor.service.service.stepexecutors;
+
+public interface StepExecutionFabric {
+
+  StepExecution getExecutionClickCss();
+
+  StepExecution getExecutionSleep();
+
+  StepExecution getExecutionXpath();
+}
