+++ src/main/java/executor/service/service/stepexecutors/selenium/StepExecutionClickCssImpl.java
@@ -1,17 +1,16 @@
-package executor.service.service.impl;
+package executor.service.service.stepexecutors.selenium;
 
-import executor.service.model.Step;
-import executor.service.service.StepExecutionClickCss;
+import executor.service.model.entity.Step;
 import org.openqa.selenium.WebDriver;
 
 public class StepExecutionClickCssImpl implements StepExecutionClickCss {
     @Override
     public String getStepAction() {
-        return null;
+        return "do smth";
     }
 
     @Override
     public void step(WebDriver webDriver, Step step) {
-
+        step.setValue("1");
     }
 }
