+++ src/main/java/executor/service/service/stepexecutors/selenium/StepExecutionClickXpath.java
@@ -1,0 +1,13 @@
+package executor.service.service.stepexecutors.selenium;
+
+import executor.service.model.entity.Step;
+import executor.service.service.stepexecutors.StepExecution;
+import org.openqa.selenium.WebDriver;
+
+public interface StepExecutionClickXpath extends StepExecution {
+
+    String getStepAction();
+
+    void step(WebDriver webDriver, Step step);
+
+}
