+++ src/main/java/executor/service/service/stepexecutors/selenium/StepExecutionSleep.java
@@ -1,0 +1,13 @@
+package executor.service.service.stepexecutors.selenium;
+
+import executor.service.model.entity.Step;
+import executor.service.service.stepexecutors.StepExecution;
+import org.openqa.selenium.WebDriver;
+
+public interface StepExecutionSleep extends StepExecution {
+
+    String getStepAction();
+
+    void step(WebDriver webDriver, Step step);
+
+}
